Update trivia_StreamlabsSystem.py
Added code to delete the question_index_map at the same time as the current_questions_list. If I didn't delete the question_index_map, there would be a variety of problems after the user switches games because the code would no longer accurately be able to map between the question in current_questions_list and master_questions_list.